+++ clients/fides-js/src/lib/consent-utils.ts
@@ -1,4 +1,3 @@
-import { isConsentOverride } from "./common-utils";
 import {
   FIDES_OVERRIDE_EXPERIENCE_LANGUAGE_VALIDATOR_MAP,
   FIDES_OVERRIDE_OPTIONS_VALIDATOR_MAP,
@@ -208,6 +207,13 @@ export const experienceIsValid = (
   }
 
   return true;
+};
+
+export const isConsentOverride = (options: FidesInitOptions) => {
+  return (
+    options.fidesConsentOverride === ConsentMethod.ACCEPT ||
+    options.fidesConsentOverride === ConsentMethod.REJECT
+  );
 };
 
 /**
